move batch mode logic
